refactor: raceOneLap() 인자받는 프로덕션 코드 작성
